Move shared examples to separate directory
