+++ spec/record_parser/sorting/by_last_name_descending_spec.rb
@@ -1,5 +1,5 @@
 require 'spec_helper'
-require 'support/shared_examples_for_sortings'
+require 'support/shared_examples/for_sortings'
 require 'support/helpers/stub_helper'
 
 module RecordParser
